:recycle: seeds: move seeds constants
