+++ api/db/seeds/data/team-certification/shared/constants.js
@@ -2,26 +2,20 @@ import {
   CLEA_COMPLEMENTARY_CERTIFICATION_ID,
   PIX_DROIT_COMPLEMENTARY_CERTIFICATION_ID,
   PIX_EDU_1ER_DEGRE_COMPLEMENTARY_CERTIFICATION_ID,
-} from '../common/complementary-certification-builder.js';
+} from '../../common/complementary-certification-builder.js';
 
 const TEAM_CERTIFICATION_OFFSET_ID = 7000;
 
 // IDS
 /// USERS
 const TEAM_CERTIFICATION_OFFSET_ID_USERS = TEAM_CERTIFICATION_OFFSET_ID + 100;
-export const PRO_ORGANIZATION_USER_ID = TEAM_CERTIFICATION_OFFSET_ID_USERS + 3;
-export const SHARED_ORGANIZATION_USER_ID = TEAM_CERTIFICATION_OFFSET_ID_USERS + 4;
-export const CERTIFIABLE_SUCCESS_USER_ID = TEAM_CERTIFICATION_OFFSET_ID_USERS + 5;
-
-/// ORGAS
-const TEAM_CERTIFICATION_OFFSET_ID_ORGAS = TEAM_CERTIFICATION_OFFSET_ID + 200;
-export const PRO_ORGANIZATION_ID = TEAM_CERTIFICATION_OFFSET_ID_ORGAS + 1;
+export const SHARED_ORGANIZATION_USER_ID = TEAM_CERTIFICATION_OFFSET_ID_USERS;
+export const CERTIFIABLE_SUCCESS_USER_ID = TEAM_CERTIFICATION_OFFSET_ID_USERS + 1;
 
 /// CERTIFICATION CENTERS
 const TEAM_CERTIFICATION_OFFSET_ID_CENTERS = TEAM_CERTIFICATION_OFFSET_ID + 300;
-export const PRO_CERTIFICATION_CENTER_ID = TEAM_CERTIFICATION_OFFSET_ID_CENTERS + 2;
-export const SHARED_CERTIFICATION_CENTER_ID = TEAM_CERTIFICATION_OFFSET_ID_CENTERS + 8;
-export const SIMPLE_SCO_CERTIFICATION_CENTER_ID = TEAM_CERTIFICATION_OFFSET_ID_CENTERS + 7;
+export const PRO_CERTIFICATION_CENTER_ID = TEAM_CERTIFICATION_OFFSET_ID_CENTERS;
+export const SCO_CERTIFICATION_CENTER_ID = TEAM_CERTIFICATION_OFFSET_ID_CENTERS + 1;
 
 /// EXTERNAL IDS
 export const SHARED_ORGANIZATION_EXTERNAL_ID = 'SHARED_CERTIFICATION_ORGANIZATION';
